Create stream handler for messages, rename files, add badger manager file
