+++ bind/node.go
@@ -1,6 +1,7 @@
 package sonr
 
 import (
+	"bufio"
 	"context"
 	"encoding/json"
 	"fmt"
@@ -14,16 +15,14 @@ import (
 // ^ Struct Management ^ //
 // Node contains all values for user
 type Node struct {
-	ctx           context.Context
-	PeerID        string
-	Host          host.Host
-	Lobby         lobby.Lobby
-	Profile       user.Profile
-	Contact       user.Contact
-	AuthStream    AuthStreamConn
-	Callback      Callback
-	DocumentPath  string
-	TemporaryPath string
+	ctx        context.Context
+	PeerID     string
+	Host       host.Host
+	Lobby      lobby.Lobby
+	Profile    user.Profile
+	Contact    user.Contact
+	AuthStream AuthStreamConn
+	Callback   Callback
 }
 
 // GetUser returns profile and contact in a map as string
@@ -53,11 +52,6 @@ func (sn *Node) SetUser(cm lobby.ConnectRequest) error {
 	// Set Contact
 	contact := user.NewContact(cm.Contact)
 	sn.Contact = contact
-
-	// Set Paths
-	sn.DocumentPath = cm.DocumentPath
-	sn.TemporaryPath = cm.TemporaryPath
-
 	return nil
 }
 
@@ -106,8 +100,12 @@ func (sn *Node) Update(data string) bool {
 
 // Invite an available peer to transfer
 func (sn *Node) Invite(id string, filePath string) bool {
-	// Get Required Data
-	peerID := sn.Lobby.GetPeerID(id)
+	// ** Get Required Data **
+	peerID, err := sn.Lobby.GetPeerID(id)
+	if err != nil {
+		fmt.Println("Search Error", err)
+		return false
+	}
 	info := user.GetInfo(sn.Profile, sn.Contact)
 
 	// Create Metadata
@@ -117,59 +115,56 @@ func (sn *Node) Invite(id string, filePath string) bool {
 		return false
 	}
 
-	// Create Request
-	request := AuthStreamMessage{
+	// ** Open a stream **
+	stream, err := sn.Host.NewStream(sn.ctx, peerID, protocol.ID("/sonr/auth"))
+
+	// Check Stream
+	if err != nil {
+		fmt.Println("Auth Stream Failed to Open ", err)
+		return false
+	}
+	// ** Create New Auth Stream **
+	// Create new Buffer
+	buffrw := bufio.NewReadWriter(bufio.NewReader(stream), bufio.NewWriter(stream))
+
+	// Create/Set Auth Stream
+	sn.AuthStream = AuthStreamConn{
+		readWriter: buffrw,
+		stream:     stream,
+		callback:   sn.Callback,
+	}
+
+	// Initialize Routine
+	go sn.AuthStream.Read()
+
+	// ** Send Invite Message **
+	err = sn.AuthStream.Write(AuthStreamMessage{
 		subject:  "Request",
 		peerInfo: info,
 		metadata: *meta,
-	}
+	})
 
-	// Convert Request to JSON String
-	msgBytes, err := json.Marshal(request)
+	// Check Error
 	if err != nil {
-		println("Error Converting Meta to JSON", err)
 		return false
 	}
 
-	// Validate then Initiate
-	if peerID != "" {
-		// open a stream, this stream will be handled by handleStream other end
-		stream, err := sn.Host.NewStream(sn.ctx, peerID, protocol.ID("/sonr/auth"))
-
-		// Check Stream
-		if err != nil {
-			fmt.Println("Auth Stream Failed to Open ", err)
-			return false
-		}
-		// Create New Auth Stream
-		sn.NewAuthStream(stream)
-
-		// Send Invite Message
-		sn.AuthStream.Send(string(msgBytes))
-
-		// Return Success
-		return true
-	}
-	return false
+	// Return Success
+	return true
 }
 
 // Accept an Invite from a Peer
 func (sn *Node) Accept() bool {
-	// Create Response
-	resp := AuthStreamMessage{
+	// Send Message
+	err := sn.AuthStream.Write(AuthStreamMessage{
 		subject:  "Response",
 		decision: true,
-	}
+	})
 
-	// Convert Response to JSON String
-	msgBytes, err := json.Marshal(resp)
+	// Check Error
 	if err != nil {
-		println("Error Converting Meta to JSON", err)
 		return false
 	}
-
-	// Send Message
-	sn.AuthStream.Send(string(msgBytes))
 
 	// Return Success
 	return true
@@ -177,21 +172,16 @@ func (sn *Node) Accept() bool {
 
 // Decline an Invite from a Peer
 func (sn *Node) Decline() bool {
-	// Create Response
-	resp := AuthStreamMessage{
+	// Send Message
+	err := sn.AuthStream.Write(AuthStreamMessage{
 		subject:  "Response",
 		decision: false,
-	}
+	})
 
-	// Convert Response to JSON String
-	msgBytes, err := json.Marshal(resp)
+	// Check Error
 	if err != nil {
-		println("Error Converting Meta to JSON", err)
 		return false
 	}
-
-	// Send Message
-	sn.AuthStream.Send(string(msgBytes))
 
 	// Return Success
 	return true
