--- bind/transport.go
@@ -1,1 +1,0 @@
-package sonr
